Split up reductions into more python files Test BalancedRed and Center reduction
